Clean up imports in sample CLI test
