+++ tests/fixtures/temp_fixtures.py
@@ -1,0 +1,1 @@
+# Placeholder fixtures for cleanup tests
